move profiling image tests to separate file
